Fix #4 when calling fragment on empty elements
VTD's [getContentFragment] [0] will return -1 if the current element is
an empty element (e.g. "<Foo/>") which would throw a
NegativeArraySizeException when using fragment. Fix this by explicitly
checking for -1 and returning an empty string to be consistent with
getting the fragment of "<Foo></Foo>".

Thanks to @borkdude for reporting this bug and helping choose the
correct behaviour.

  [0]: https://vtd-xml.sourceforge.io/javadoc/com/ximpleware/VTDNav.html#getContentFragment--